requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
requests==2.31.0 
python-dotenv==1.1.0
langgraph==0.3.24
pydantic==2.11.2
beautifulsoup4==4.13.3
langchain-groq==0.2.1
streamlit==1.32.0
# Dependências adicionais para garantir compatibilidade
typing-extensions>=4.5.0
httpx==0.27.0